requirements.txt
altgraph==0.17
app-version==1.0.1
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
httplib2==0.9.2
idna==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
notify==0.3.1
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pexpect==4.2.1
Pillow==7.2.0
protobuf==3.0.0
PyAudio==0.2.11
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyInstaller==3.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyperclip==1.8.0
pyRFC3339==1.0
pyshorteners==1.0.1
PySocks==1.7.1
pytesseract==0.3.5
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.14.0
pyttsx3==2.90
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
SpeechRecognition==3.8.1
system-service==0.3
systemd-python==234
tweepy==3.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
validate-email==1.3
vboxapi==1.0
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
